check if avatar null

+++ resources/Uploads.py
@@ -69,7 +69,8 @@ class UserAvatar(Resource):
         url = "https://%s.s3-%s.amazonaws.com/%s" % (bucket_name, location, key)
 
         # delete old pfp
-        s3.delete_object(Bucket = BUCKET_NAME,Key=profile.avatar.split("/")[-1])
+        if(profile.avatar!= None):
+            s3.delete_object(Bucket = BUCKET_NAME,Key=profile.avatar.split("/")[-1])
 
         # set pfp
         profile.avatar = url
